+++ gen/com/intellij/idea/plugin/hybris/impex/psi/ImpexUserRightsMultiValue.java
@@ -25,7 +25,7 @@ import java.util.List;
 import org.jetbrains.annotations.*;
 import com.intellij.psi.PsiElement;
 
-public interface ImpexUserRightsValue extends PsiElement {
+public interface ImpexUserRightsMultiValue extends ImpexUserRightsValue {
 
   @Nullable
   ImpexUserRightsHeaderParameter getHeaderParameter();