+++ src/com/intellij/idea/plugin/hybris/impex/psi/ImpexUserRightsValue.java
@@ -21,25 +21,15 @@
  */
 package com.intellij.idea.plugin.hybris.impex.psi;
 
-import java.util.List;
-import org.jetbrains.annotations.*;
 import com.intellij.psi.PsiElement;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
-public interface ImpexUserRightsFirstValueGroup extends PsiElement {
+import java.util.List;
 
-  @Nullable
-  ImpexUserRightsAttributeValue getUserRightsAttributeValue();
-
-  @Nullable
-  ImpexUserRightsMultiValue getUserRightsMultiValue();
+public interface ImpexUserRightsValue extends PsiElement {
 
   @Nullable
-  ImpexUserRightsPermissionValue getUserRightsPermissionValue();
-
-  @Nullable
-  ImpexUserRightsSingleValue getUserRightsSingleValue();
-
-  @Nullable
-  ImpexValueLine getValueLine();
+  ImpexUserRightsHeaderParameter getHeaderParameter();
 
 }